pyproject.toml
[tool.poetry]
name = "koi"
version = "0.5.0"
description = "Authenticated banking copilot MVP"
authors = ["David Llorente <davidllorenteraposo@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.136.0"
uvicorn = "^0.44.0"
pydantic-settings = "^2.13.1"
python-dotenv = "^1.2.2"
email-validator = "^2.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^5.0.0"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
pyjwt = "^2.12.1"
sqlmodel = "^0.0.38"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.3"
ruff = "^0.15.11"
black = "^26.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
